+++ src/org/traccar/config/ConfigSuffix.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2019 Anton Tananaev (anton@traccar.org)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.traccar.config;
+
+public class ConfigSuffix extends ConfigKey {
+
+    ConfigSuffix(String key, Class clazz, String description) {
+        super(key, clazz, description);
+    }
+
+    public ConfigKey withPrefix(String prefix) {
+        return new ConfigKey(prefix + getKey(), getValueClass(), getDescription());
+    }
+
+}
