+++ src/org/traccar/handler/FilterHandler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.traccar.processing;
+package org.traccar.handler;
 
 import io.netty.channel.ChannelHandler;
 import org.slf4j.Logger;
